requirements.txt
flask
flask-socketio
confluent-kafka
google-cloud-secret-manager
google-auth
simple-websocket
google-cloud-aiplatform